Create setBackground.js
Select a random background image from the config file's list of backgrounds.
Get a reference to the body element.
Set the style attribute and add the new background image.
The baseUrl function gets the correct context for the background images to be loaded from either the web server root '/www' or the local path "C://.."